fixed shapes in grid plots
