--- bloptools/functions.py
@@ -1,5 +1,0 @@
-import numpy as np
-
-
-def sigmoid(x):
-    return 1 / (1 + np.exp(-x))
